+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### Features
+
+- [BREAKING] Added CLI for the transaction prover services both the workers and the proxy (#955).
+
+### Fixes
+
+- Fixed `AccountId::new_with_type_and_mode()` (#958). 
+
 ## 0.6.0 (2024-11-05)
 
 ### Features